Improving README
The README does a great job of asisting incoming developers, but does not effectively describe
the project or provide compelling reasons to contribute.  I have added a link to the home page
as well as a short description of its purpose and achievements to date based on a conversation
with the developers.  Please feel free to recommend any changes to improve wording or content!